requirements.txt
fastapi
uvicorn
pandas
python-dotenv
pyyaml
gunicorn